// ReactApp/src/screens/ScheduleScreen/types.ts
export type ServiceProps = {
    id: number;
    text: string;
}

export type TherapistProps = {
    id: number;
    text: string;
}

export  type ValuesTypes = {
    firstName: string;
    lastName: string;
    email: string;
    childName: string;
    selectedService: string;
    selectedTherapist: string;
    selectedHour: string;
}